Extract terminal host overlay support
