+++ clients/apple/ARCHITECTURE.md
@@ -124,12 +124,10 @@ device support was not required for this validation.
 ## Remaining Architecture Debt
 
 `check-architecture-maintainability.sh` currently completes in report mode with
-four known warnings:
+three known warnings:
 
 - `ShellHostController.swift` remains large pending additional controller,
   store, and projection splits.
-- `TerminalHostView.swift` remains large pending additional terminal-host
-  collaborator extraction.
 - `Views/Console/ContentView.swift` remains large and imports AppKit because the
   legacy/mobile console path is isolated but not fully decomposed.
 The current architecture gate intentionally keeps those warnings non-blocking
